requirements.txt
h3cy==0.0.3
google-api-core==1.15.0
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-cloud-bigquery[bqstorage,pandas]==1.23.1
google-cloud-core==1.1.0
google-cloud-storage==1.23.0
googleapis-common-protos==1.6.0
six==1.13.0
google-resumable-media==0.5.0